"Fix" default audio stream not being used. Not really a proper fix, I think the original intention of this conditional was to set ai to the first multichannel (more than 2 audio channels) audio stream index and change ai to the first stereo if one is found.
Without this change, if more than one audio stream is stereo, the last one
is used, when we want the first one to be used.

The proper fix would be to check if ai is already set to a stereo stream,
which would require tracking the channels for the last audio index.